refactor: extract train & sample logic into operations
